新增IDE功能模块和文件处理相关方法
- 添加IDE资源管理视图和文件处理视图
- 实现资源的创建、删除、重命名等操作
-增加文件内容读取和判断文件类型的方法
- 优化Spider模型的资源路径处理逻辑
- 新增异常处理器以处理数据被使用中的操作
- 引入chardet库以支持文件编码检测

+++ apps/Spider/models.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 from uuid import uuid4
 
 from django.db import models
@@ -19,5 +20,6 @@ class SpiderModel(models.Model):
     updateTime = models.DateTimeField(auto_now=True, verbose_name='更新时间')
 
     def delete(self, using=None, keep_parents=False):
-        shutil.rmtree(self.resources)
+        if Path(self.resources).exists():
+            shutil.rmtree(self.resources)
         return super().delete(using, keep_parents)
